requirements.txt
Flask==1.1.1
Flask-JWT==0.3.2
requests==2.22.0
psycopg2-binary==2.8.4
tenacity==6.0.0